feat: add routing with error page
