+++ src/pages/ErrorPage.tsx
@@ -1,0 +1,7 @@
+function ErrorPage() {
+  return ( <h3>
+    Sorry the requested route is not found.
+  </h3> );
+}
+
+export default ErrorPage;
