+++ src/pages/GreetingPage.tsx
@@ -1,0 +1,7 @@
+function GreetingPage() {
+  return ( <h1>
+    Hello, one day this page will be more popular, but not today...
+  </h1> );
+}
+
+export default GreetingPage;
