+++ src/pages/MainPage.tsx
@@ -1,0 +1,7 @@
+import MainComponent from "../components/MainComponent";
+
+function MainPage() {
+  return ( <MainComponent/> );
+}
+
+export default MainPage;
